# nhs-headless-user-readiness-impl/rest-openapi.yaml
components:
    schemas:
        Question:
            type: object
            required:
                - id
                - name
                - index
                - selectedAnswer
                - answers
            properties:
                id:
                    type: integer
                    format: int64
                name:
                    type: string
                index:
                    type: integer
                    format: int32
                selectedAnswer:
                    type: string
                options:
                    type: array
                    items:
                        type: string
info:
    description: "NHS Talent User Readiness"
    license:
        name: "Apache 2.0"
        url: "http://www.apache.org/licenses/LICENSE-2.0.html"
    title: "NHS Talent"
    version: "1.0.0"
openapi: 3.0.1
paths:
    "/user-profile/questionnaire":
        get:
            description: Returns the whole user readiness questionnaire
            operationId: getUserProfileQuestionnairePage
            responses:
                200:
                    content:
                        application/json:
                            schema:
                                type: "array"
                                items:
                                    $ref: "#/components/schemas/Question"
                        application/xml:
                            schema:
                                type: "array"
                                items:
                                    $ref: "#/components/schemas/Question"
                    description: "success"
            tags: ["Question"]
    "/user-profile/questionnaire-intro":
        get:
            description: Returns the intro content for the questionnaire
            operationId: getUserProfileQuestionnaireIntro
            responses:
                200:
                    content:
                        application/json:
                            schema:
                                type: "string"
                        application/xml:
                            schema:
                                type: "string"
                    description: "success"
            tags: ["Question"]
    "/user-profile/question/{questionId}":
        get:
            description: Returns a single question
            operationId: getUserProfileQuestion
            parameters:
                - in: path
                  name: questionId
                  required: true
                  schema:
                      format: int64
                      type: integer
            responses:
                200:
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/Question"
                        application/xml:
                            schema:
                                $ref: "#/components/schemas/Question"
                    description: "success"
            tags: ["Question"]
        put:
            description: Updates the answer to a single question
            operationId: putUserProfileQuestion
            parameters:
                - in: path
                  name: questionId
                  required: true
                  schema:
                      format: int64
                      type: integer
            requestBody:
                content:
                    text/plain:
                        schema:
                            type: string
            responses:
                200:
                    content:
                        {}
                    description: "success"
            tags: ["Question"]
        delete:
            description: Deletes the answer for a single question
            operationId: deleteUserProfileQuestion
            parameters:
                - in: path
                  name: questionId
                  required: true
                  schema:
                      format: int64
                      type: integer
            responses:
                200:
                    content:
                        {}
                    description: "success"
            tags: ["Question"]